+++ src/components/ButtonGroup/facade/Facade.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+class Facade extends React.PureComponent {}
+
+export default Facade