+++ src/components/ButtonGroup/styles/index.js
@@ -1,0 +1,1 @@
+export StyledButtonGroup from './StyledButtonGroup'