+++ src/components/ButtonGroup/theme/index.js
@@ -1,0 +1,9 @@
+export default ({ borders, colors }) => ({
+  borders: {
+    borderWidth: borders.widths.bw2,
+    borderRadius: borders.radius.br3,
+  },
+  colors: {
+    border: colors.extraDarkSnow,
+  },
+})