+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Please beware that all version pre 1.0.0 can have major and breaking changes in 
 - Added support for custom renderers into Breadcrumb [@Atrakeur](https://github.com/atrakeur)
 - Added support for custom renderers into Menu [@Atrakeur](https://github.com/atrakeur)
 - Added Table component [@Atrakeur](https://github.com/atrakeur)
+- Added ButtonGroup component [@Atrakeur](https://github.com/atrakeur)
 
 ### Changed